Improvements to top-level server error reporting (#5349)
* For MultiErrors reported from task groups, don't lose
   chained exceptions
 * When an InternalServerError kills the server, produce
   the "please file a bug" message
