+++ edb/common/taskgroup.py
@@ -22,9 +22,9 @@ from __future__ import annotations
 import asyncio
 import itertools
 import sys
-import textwrap
-import traceback
 import types
+
+from . import multi_error
 
 
 class TaskGroup:
@@ -286,29 +286,7 @@ class TaskGroup:
             self._parent_task.cancel()
 
 
-class MultiError(Exception):
-
-    def __init__(self, msg, *args, errors=()):
-        if errors:
-            types = set(type(e).__name__ for e in errors)
-            msg = f'{msg}; {len(errors)} sub errors: ({", ".join(types)})'
-            for er in errors:
-                msg += f'\n + {type(er).__name__}: {er}'
-                if er.__traceback__:
-                    er_tb = ''.join(traceback.format_tb(er.__traceback__))
-                    er_tb = textwrap.indent(er_tb, ' | ')
-                    msg += f'\n{er_tb}\n'
-        super().__init__(msg, *args)
-        self.__errors__ = tuple(errors)
-
-    def get_error_types(self):
-        return {type(e) for e in self.__errors__}
-
-    def __reduce__(self):
-        return (type(self), (self.args,), {'__errors__': self.__errors__})
-
-
-class TaskGroupError(MultiError):
+class TaskGroupError(multi_error.MultiError):
     pass
